refactor: consolidate contact modal logic into ContactModalForm component and update related interfaces
