+++ frontend/src/community/crm/types/CommonTypes.ts
@@ -103,7 +103,7 @@ export interface CrmOwnersResponseType {
   totalPages: number;
 }
 
-export interface CrmContactAddFormTypes {
+export interface CrmContactFormValues {
   name: string;
   email: string;
   contactNumber: string;
@@ -119,16 +119,8 @@ export interface CrmContactCreatePayload {
   ownerId?: number;
 }
 
-export interface CrmContactEditFormTypes {
-  name: string;
-  email: string;
-  contactNumber: string;
-  companyId: number | null;
-  ownerId: number | null;
-}
-
 export interface EditContactPayload extends CrmContactCreatePayload {
-  id?: number;
+  id: number;
 }
 
 export interface CrmContactMetricsType {
